Recorded Journey Data Logs
Journeys of varied lengths. Incident journey named, the incident occurs at 22:40:14. DriveLog203040mphBraking.csv is a riding controlled test set with acceleration up to 20mph, then harder breaking and a pause before the next. Up to 40mph. A roundabout is included and recognisable by the extended pause after this manouver, before an acceleration to final 40mph. Stationary Data was recorded on a hard surface in a building with no/minimal environmental disturbences. Stationary Bike Noise was recorded on a running but static motorcycle.  to gather engine vibration data.